React/AIFlashcardWithReact/FlashCardGeneratorAI-main/requirements.txt
Flask
pypdf
python-pptx
google-generativeai
